Switched to VQE example and reformatted into md file
I went back to the VQE example I started with and managed to get it to run with Mitiq.
After working with the variational classifier example a bit more, I realize the runtime is too long for a tutorial example. I think it’s because I’m using a simulator and the circuit depth is too large to run efficiently.

I still need to experiment with the noise scaling and extrapolation to get the VQE example to converge, but at least it runs more efficiently than the classifier example.