avoid using `ref` role when not needed
the `doc` role is more suited for the purposes here of linking to pages.
If one wishes to link to a specific section, the `ref` role would be the
way to go. Here we were mostly only using `ref` superficially to link to
headers. If that's the case linking to the document itself makes most
sense.