Speed up tests using mocks (#2126)
* fix typehint

this function always gets called with a single PauliString object, and
it wouldn't make sense to call with a list/sequence

* mock `execute_with_qse` for non-direct calls

these two tests do not need to run any of the computation done in the
`test_execute_with_qse` since that test already covers that! no need to
check performance either as, again, the previous test covers that. these
tests exist only to ensure the two alternate workflows run.

* use a fixture to simplify setup in files

* remove unnecessary typehint

* move `final_executor` definition above function

seems very strange to me to have it defined after the function that uses
it. It's valid, but harder to understand IMO.

* simplify variable naming

* DRY up code

* use `numpy.full` instead of multiplication

* clean up Hamiltonian overlap func

* consolidate overlap matrix functions

* use new `cache` variable name

* add missing `assert` statement

* clean up comments

* mock expensive executors and fix cost estimator

mocking this unearthed a bug with the cost estimator. Since for each
problem/circuit, we compute the noisy value in addition to the EM, we
were actually undercounting the estimation by `num_circuits`

* use caps to indicate objecs are global/static

closes #2015 (discussion)

* format